# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/guest": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "summary": "Get all guests",
                "description": "List of guests",
                "operationId": "bdd3e88cba3e6bcf18a3beab32adff44",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Guest"
                ],
                "summary": "Create a new guest",
                "description": "Create a new guest with the provided details",
                "operationId": "174c3e57f2db9deb6567d79efd972a70",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "lastName",
                                    "dniPassport",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Guest Name"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Guest Last Name"
                                    },
                                    "dniPassport": {
                                        "type": "string",
                                        "example": "25588978L"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "guest@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/guest/{id}": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "summary": "Get a specific guest",
                "description": "Retrieve a specific guest by ID",
                "operationId": "d85743487481a82c98b04b8b0f09df98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Guest"
                ],
                "summary": "Update an existing guest",
                "description": "Update an existing guest by ID",
                "operationId": "d10c7f583afe5a0e1cdb1b511a59faf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "lastName",
                                    "dniPassport",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Guest Name"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Updated Guest Last Name"
                                    },
                                    "dniPassport": {
                                        "type": "string",
                                        "example": "89653691D"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "updated-guest-email@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest updated successfully"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guest"
                ],
                "summary": "Delete a specific guest",
                "description": "Delete a specific guest by ID",
                "operationId": "883beca2e205c4f94d5d8768c6af3edb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guest deleted successfully"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        },
        "/api/guest/search/{name}": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "summary": "Search guests by name",
                "description": "Search guests by name",
                "operationId": "b88cc0703773187958b5181e1f26b133",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the guest",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/guest/{id}/rooms": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "summary": "Get rooms of a specific guest",
                "description": "Get rooms of a specific guest by ID",
                "operationId": "5db7f10ac82052b461280c6d8fa653ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        },
        "/api/guest/{guest}/rooms": {
            "post": {
                "tags": [
                    "Guest"
                ],
                "summary": "Add rooms to a specific guest",
                "description": "Add rooms to a specific guest by ID",
                "operationId": "96a548c79fd3e92d066381b9c8d98825",
                "parameters": [
                    {
                        "name": "guest",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "guest_id",
                                    "room_id",
                                    "checkInDate",
                                    "checkOutDate"
                                ],
                                "properties": {
                                    "guest_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "checkInDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-10-01"
                                    },
                                    "checkOutDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-10-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rooms added successfully"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guest"
                ],
                "summary": "Remove rooms from a specific guest",
                "description": "Remove rooms from a specific guest by ID",
                "operationId": "4c3027fea29181c8a7e62ba204a4e6d3",
                "parameters": [
                    {
                        "name": "guest",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "guest_id",
                                    "room_id",
                                    "checkInDate",
                                    "checkOutDate"
                                ],
                                "properties": {
                                    "guest_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "checkInDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-10-01"
                                    },
                                    "checkOutDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-10-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "rooms removed successfully"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        },
        "/api/guest/{id}/services": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "summary": "Get services of a specific guest",
                "description": "Get services of a specific guest by ID",
                "operationId": "a6dde57f2f45db9a428f17822fd23d9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        },
...
    },
...
}